[project_utils,+test_project_utils]simplify + test
The project_utils is cleaned up by removing some os.chdir logic that set
the working directory till exactly the same directory. So this did not
do anything.
Also the dependency to dataclass_csv is removed. This is a package that
has not been updated since 2021 and is not widely used. Instead, we use
the csv package that is standard included in python, which ensures that
it stays up to date and this removes a dependency.
Some comments are added to easier distinguish between code that is
specific for Cloudina and the general logic.

There is also a test created for the find_project function, both to test
if values get assigned to the project class correctly, and to see if the
error gets raised when the project does not exist.
It is not tested if the info messages are provided when the project
exists.