[db_utils,project,test_db_utils] improve API design around db creation
Instead of doing double work around first creating the DB, and then
requesting a connection. This considerably simplifies the tests, and
avoids a potential issue when create_connection is called before the
DB has actually been created.

In the tests we also add fixtures, since we want to run the tests on a
clean DB every time. This fixture creates and removes the DB on every
test.