make project specific utils_path a project property
Instead of having many places in the codebase where we check if a
project belongs to a certain project, and in that case do something, we
want this logic to all be present in the project_list.csv, which is the
project config file.
To do that, this commit adds an extra column with the utils_path.
Importlib.import_module can import this module dynamically and retrieve
the function from there that is needed.

This we want to do with all the places where a specific project is
named. This logic should all live in one place, the project_config which
is the Project class, created via the project_list.csv.