rewrite process_test_csv and write test
This commit simplifies the code in process_tes_csv. The logic that there
was for the movies was undone by itself. It added the site_id column but
then later removed it by only selecting the columns that are in schema
and there site_id is not present. So this part is removed.

I also removed the function in movie_utils and put it inside the
process_test_csv since it is the only place where it is used. This
project specific logic we would like to live somewhere else, but for now
it can stay where it is.

Next there are tests added for all different scenarios of using this
function.

To keep testing at the lowest level possible, we want the test for
test_table_for_none to be its own test function instead of being part of
test_process_test_csv