[project.py] usage general_utils.import_modules
The ProjectProcessor does not use the self.import_modules at all, so we
can remove this here.
In the other places we do not need to initialise it first by calling the
function empty and returning an empty dict. We can directly call the
function with the modules it should import.