[project.py] remove Annotator
Annotator is not used anywhere except for insdie the src folder. I
cannot find where the src folder is used. If it is not used, I can
remove that too and squash it with this commit. If it ends up to be
used, I should not remove the Annotator function, but put it inside the
src folder instead.