[project_utils, widgets.py] reuse the same code
Both functions had logic to see if the projects_list.csv file exist and
otherwise download it. I now broke this out to its own function and used
this in both. In that way the widget function can also just stay the
widget.

Another change is the definition of the path, instead of setting it
absolutely, it is now defined from the kso_utils folder by using where
the file with the function is located. In this way, it should be
possible to run this utils function from all different scripts and
notebooks. Instead of that you get a path error because the notebook
looks for the file compared to its own path.