Check_metadata: remove functionality to change metadata
We do want to keep that we can automatically check if all movies exist
that are described in the movies.csv, and that we add the fsp, start and
stop time, etc. But we do not want to keep the functionality where we
correct for spelling mistakes in the csv's themselves.

This never really worked due to folder permissions issues and is easy to
do manual. So that part is removed here.