remove all 'if test' statements from old notebook testing
These are left overs from a previous attempt to create tests. These
tests were not maintained and were ignored when code was changed, so
they are not up to date. Therefore, they were removed in another commit
and this commit removes some of its left overs in the code itself.